vps/requirements.txt
supabase
python-dotenv
feedparser
